Renaming multi_seq_align.py file and removing old version
